PiDataProcessor/requirements.txt
firebase-admin~=6.6.0
google-auth
requests~=2.32.3
Pillow~=11.0.0
paho-mqtt~=2.1.0
numpy~=2.1.3
protobuf~=5.29.0